src/test/java/be/kc/persondata/controller/PersonDataControllerTest.java
package be.kc.persondata.controller;

import be.kc.persondata.model.PersonData;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.boot.test.web.client.TestRestTemplate;
import org.springframework.boot.web.server.LocalServerPort;

import static org.assertj.core.api.Assertions.assertThat;
import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;

@SpringBootTest(webEnvironment = RANDOM_PORT)
class PersonDataControllerTest {

    @Autowired
    PersonDataController controller;

    @LocalServerPort
    private int port;

    @Autowired
    private TestRestTemplate restTemplate;

    @Test
    void testLoad() throws Exception {
        assertThat(restTemplate.getForEntity("http://localhost:" + port + "/load", PersonData[].class))
                .isNotNull();
    }
}